Implemented GraphQL-based API gateway

+++ build-and-test-all.sh
@@ -1,14 +1,37 @@
 #! /bin/bash -e
 
+KEEP_RUNNING=
+ASSEMBLE_ONLY=
+
+while [ ! -z "$*" ] ; do
+  case $1 in
+    "--keep-running" )
+      KEEP_RUNNING=yes
+      ;;
+    "--assemble-only" )
+      ASSEMBLE_ONLY=yes
+      ;;
+    "--help" )
+      echo ./build-and-test-all.sh --keep-running --assemble-only
+      exit 0
+      ;;
+  esac
+  shift
+done
+
+echo KEEP_RUNNING=$KEEP_RUNNING
+
 . ./set-env.sh
 
 initializeDynamoDB() {
-    echo preparing dynamodblocal table data
-    cd dynamodblocal-init
-    ./create-dynamodb-tables.sh
-    cd ..
-    echo data is prepared
+./initialize-dynamodb.sh
 }
+
+# TODO Temporarily
+
+./build-contracts.sh
+
+./gradlew testClasses
 
 docker-compose down -v
 docker-compose up -d --build dynamodblocal mysql
@@ -21,25 +44,32 @@ initializeDynamoDB
 
 docker-compose up -d --build eventuate-local-cdc-service tram-cdc-service
 
-# TODO Temporarily
 
-./build-contracts.sh
+if [ -z "$ASSEMBLE_ONLY" ] ; then
 
-./gradlew -x :ftgo-end-to-end-tests:test $* build
+  ./gradlew -x :ftgo-end-to-end-tests:test $* build
 
-docker-compose build
+  docker-compose build
 
-./gradlew $* integrationTest
+  ./gradlew $* integrationTest
 
 
-# Component tests need to use the per-service database schema
+  # Component tests need to use the per-service database schema
 
-SPRING_DATASOURCE_URL=jdbc:mysql://${DOCKER_HOST_IP?}/ftgoorderservice ./gradlew :ftgo-order-service:cleanComponentTest :ftgo-order-service:componentTest
+  SPRING_DATASOURCE_URL=jdbc:mysql://${DOCKER_HOST_IP?}/ftgoorderservice ./gradlew :ftgo-order-service:cleanComponentTest :ftgo-order-service:componentTest
 
-# Reset the DB/messages
+  # Reset the DB/messages
 
-docker-compose down -v
-docker-compose up -d
+  docker-compose down -v
+  docker-compose up -d
+
+
+else
+
+  ./gradlew $* assemble
+  docker-compose up -d --build
+
+fi
 
 ./wait-for-mysql.sh
 
@@ -47,10 +77,16 @@ echo mysql is started
 
 initializeDynamoDB
 
-date
+./wait-for-mysql.sh
+
+echo mysql is started
 
 ./wait-for-services.sh
 
 ./gradlew :ftgo-end-to-end-tests:cleanTest :ftgo-end-to-end-tests:test
 
-docker-compose down -v
+./run-graphql-api-gateway-tests.sh
+
+if [ -z "$KEEP_RUNNING" ] ; then
+  docker-compose down -v
+fi
